pyproject.toml
[project]
name = "offal"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Matthew White", email = "matthew.r.j.white@gmail.com" }
]
dependencies = [
    "gitpython>=3.1.43",
    "typer>=0.12.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"offal" = "offal:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[pytest]
mock_use_standalone_module = true

[tool.ruff]
line-length = 120
src = [
    "src",
    "tests",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.2",
    "pytest-mock>=3.14.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/offal", "src/offal/py.typed"]
